chore: refactor pysimplegui controllers
