+++ movie_pipeline_segments_validator/views/segments_timeline.py
@@ -1,7 +1,7 @@
 from typing import Any
 import PySimpleGUI as sg
 
-from ..controllers.segments_timeline_controller import draw_current_position_indicator, draw_segments, resize_timeline, update_current_position_indicator_position, update_selected_segment
+from ..adapters.pysimplegui.segments_timeline_controller import draw_current_position_indicator, draw_segments, resize_timeline, update_current_position_indicator_position, update_selected_segment
 from ..domain.widget import WidgetEvent, WidgetKey
 from ..domain.context import TimelineContext
